Fix signed/unsigned compare warning
The underlying type of DC::EnumType is implementation dependent.
It might be better to construct a vtkDICOMTag from DC::EnumType,
but that might impact performance.  Actually the use of DC::EnumType
for an array of tags is not ideal, for two reasons:
1) this might actually be a 64-bit type when only 32 bits are needed,
2) there are valid tags that are not DC::EnumType members.

+++ Source/vtkDICOMGenerator.cxx
@@ -1051,7 +1051,7 @@ bool vtkDICOMGenerator::CopyAttributes(
         const DC::EnumType *tags = blacklist;
         while (*tags != DC::ItemDelimitationItem)
         {
-          if (*tags++ == tag.GetKey())
+          if (static_cast<unsigned int>(*tags++) == tag.GetKey())
           {
             blacklisted = true;
             break;
